+++ src/fsharp/vs/service.fs
@@ -412,7 +412,7 @@ type TypeCheckInfo
            // This is a name resolution environment to use if no better match can be found.
            sFallback:Nameres.NameResolutionEnv,
            loadClosure : LoadClosure option,
-           startOp:(unit->unit)->unit,
+           reactorOps : IReactorOperations,
            checkAlive : (unit -> bool),
            textSnapshotInfo:obj option) = 
 
@@ -960,7 +960,7 @@ type TypeCheckInfo
                 | Some(items,denv,m) -> 
                     let items = items |> filterIntellisenseCompletionsBasedOnParseContext (parseResultsOpt |> Option.bind (fun x -> x.ParseTree)) (mkPos line colAtEndOfNamesAndResidue)
                     let items = if isInterfaceFile then items |> List.filter IsValidSignatureFileItem else items
-                    DeclarationSet.Create(infoReader,m,denv,items,startOp,checkAlive))
+                    DeclarationSet.Create(infoReader,m,denv,items,reactorOps,checkAlive))
             (fun msg -> DeclarationSet.Error msg)
             
     member scope.GetReferenceResolutionToolTipText(line,col) : ToolTipText = 
@@ -1382,7 +1382,7 @@ module internal Parser =
            loadClosure: LoadClosure option,
            // These are the errors and warnings seen by the background compiler for the entire antecedant 
            backgroundErrors: (PhasedError * bool) list,    
-           startOp: (unit->unit)->unit,
+           reactorOps: IReactorOperations,
            // Used by 'DeclarationSet' to check the IncrementalBuilder is still alive.
            checkAlive : (unit -> bool),
            isResultObsolete: unit->bool,
@@ -1508,7 +1508,7 @@ module internal Parser =
                                 sink.GetTcResolutions(), 
                                 tcEnvAtEnd.NameEnv,
                                 loadClosure,
-                                startOp,
+                                reactorOps,
                                 checkAlive,
                                 textSnapshotInfo)     
                 errors, TypeCheckAborted.No scope
@@ -1563,11 +1563,9 @@ type ProjectOptions =
             sb.ToString()
         sprintf "ProjectOptions(%s)\n  Files:\n%s  Options: %s" this.ProjectFileName files options
  
-type IReactorOperations = 
-    abstract RunAsyncOp : (unit -> 'T) -> Async<'T>
-    abstract StartAsyncOp: Reactor.Operation -> unit
 
-type [<Sealed>] ProjectContext(assemblies: FSharpAssembly list) =
+[<Sealed>] 
+type ProjectContext(assemblies: FSharpAssembly list) =
 
     /// Get the assemblies referenced
     member __.GetReferencedAssemblies() = assemblies
@@ -1995,7 +1993,7 @@ type BackgroundCompiler() as self =
                     // Run the type checking.
                     let tcErrors, tcFileResult = 
                         Parser.TypeCheckOneFile(parseResults,source,filename,options.ProjectFileName,tcConfig,tcGlobals,tcImports,  tcPriorState,
-                                                loadClosure,backgroundErrors,reactor.StartAsyncOp,(fun () -> builder.IsAlive),isResultObsolete,textSnapshotInfo)
+                                                loadClosure,backgroundErrors,reactorOps,(fun () -> builder.IsAlive),isResultObsolete,textSnapshotInfo)
 
                     Some(MakeCheckFileAnswer(tcFileResult, options, builder, creationErrors, parseResults.Errors, tcErrors))
                 | None -> None
@@ -2012,7 +2010,7 @@ type BackgroundCompiler() as self =
             let loadClosure = scriptClosure.TryGet options 
             let tcErrors, tcFileResult = 
                 Parser.TypeCheckOneFile(parseResults,source,filename,options.ProjectFileName,tcConfig,tcGlobals,tcImports,  tcPriorState,
-                                        loadClosure,priorErrors,reactor.StartAsyncOp,(fun () -> builder.IsAlive),isResultObsolete,textSnapshotInfo)
+                                        loadClosure,priorErrors,reactorOps,(fun () -> builder.IsAlive),isResultObsolete,textSnapshotInfo)
             MakeCheckFileAnswer(tcFileResult, options, builder, creationErrors, parseResults.Errors, tcErrors)
 
     /// Parses the source file and returns untyped AST
@@ -2043,7 +2041,7 @@ type BackgroundCompiler() as self =
             let loadClosure = scriptClosure.TryGet options 
             let tcErrors, tcFileResult = 
                 Parser.TypeCheckOneFile(parseResults,source,filename,options.ProjectFileName,tcConfig,tcGlobals,tcImports,  tcPriorState,
-                                        loadClosure,priorErrors,reactor.StartAsyncOp,(fun () -> builder.IsAlive),isResultObsolete,textSnapshotInfo)
+                                        loadClosure,priorErrors,reactorOps,(fun () -> builder.IsAlive),isResultObsolete,textSnapshotInfo)
             let checkAnswer = MakeCheckFileAnswer(tcFileResult, options, builder, creationErrors, parseResults.Errors, tcErrors)
             parseResults, checkAnswer, false
 
@@ -2067,7 +2065,7 @@ type BackgroundCompiler() as self =
                 let scope = 
                     TypeCheckInfo(tcConfig, tcGlobals, tcState.PartialAssemblySignature, tcState.Ccu, tcImports,
                                   options.ProjectFileName, filename, List.last tcResolutions, tcEnvAtEnd.NameEnv,
-                                  loadClosure, reactorOps.StartAsyncOp, (fun () -> builder.IsAlive), None)     
+                                  loadClosure, reactorOps, (fun () -> builder.IsAlive), None)     
                 let typedResults = MakeCheckFileResults(options, builder, scope, creationErrors, parseResults.Errors, tcErrors)
                 (parseResults, typedResults)
 
@@ -2416,7 +2414,7 @@ type FsiInteractiveChecker(reactorOps: IReactorOperations, tcConfig, tcGlobals, 
         let backgroundErrors = []
         let tcErrors, tcFileResult = 
             Parser.TypeCheckOneFile(parseResults,source,mainInputFileName,"project",tcConfig,tcGlobals,tcImports,  tcState,
-                                    loadClosure,backgroundErrors,reactorOps.StartAsyncOp,(fun () -> true),(fun _ -> false),None)
+                                    loadClosure,backgroundErrors,reactorOps,(fun () -> true),(fun _ -> false),None)
 
         match tcFileResult with 
         | Parser.TypeCheckAborted.No scope ->